Moving place revamp test to Base DC only.
